Moving test to own test_generation_utils, use of lighter models for testing
